Simplify app and setup GitHub CI/CD
The changes removed styles from `globals.css` and `page.module.css` simplifying the layout of the app. The `README.md` was also updated to make it specific to the project. A new GitHub workflow file was added for CI/CD pipeline, it will trigger on main branch and use Node.js 16.x and pnpm for building and deploying the app. The `next.config.js` was updated for setup base path for production environment.

These changes make the project easier to manage and deploy while also improving the UI consistency.