Add new components and refactor Home to use these
Added new components: New, Popular, Video, Article for displaying data. Refactored Home component to use these newly created components and displaying a hardcoded list of video and article data.

The aim is to simplify the structure and improve the versatility and reusability in the code. Also, to demonstrate how data types can be generalised and reused across components. The update has profoundly simplified the `page.tsx` structure, and now it is more adaptive to additional data types. We can add any list item, and it can be an article or a video.